Make witness size generic (#49)
Change witness entries from i64 integers to Int<N> where N is arbitrarily large.
i.e. width of integers is 64 * N bits

Encoding matrix entries are fixed to be double the width of the witness entries.
The code entries are 4 * width of the witness entries
The Linear combination of codes has entries 8 * width of witness entries