Integrate Merkle Tree implementation from Plonky3 (#140)
Reworked Merkle Tree approach used in Zinc.

Instead of a makeshift implementation, rely on the one designed by https://github.com/Plonky3/Plonky3 - just like https://github.com/tcoratger/whir-p3 does.

This particular MT is designed to commit to matrices, each matrix row becoming an MT leaf that is opened together. This is very convenient for Zinc since we need to open entire columns. For that, we transpose a matrix before committing.

Most changes in this PR are trivial due to required implementation of `Copy` trait for `Integer`. Main changes are:
* Use `Plonky3`'s MT implementation to commit to entire u-hat matrix
* There's now one MT root hash for a matrix, instead of a one per-row
* This also makes proof size much smaller due to only one Merkle proof being needed per opening
* Hash type has been wrapped in `MtHash` for ease of change
* Fixed an issue in `verify_testing` where the result of `verify_column` was silently ignored
* In addition, `encode_rows` and matrix transposal is made using `MaybeUninit` and a minimal amount of unsafe code, yielding ~10% speedup of a `commit` operation.